gcp/functions/hybrid-scanner/requirements.txt
functions-framework==3.5.0
google-cloud-storage==2.14.0
python-nmap==0.7.1
requests==2.31.0
